Move fake-api into proxy and extract proxy-invoke
