+++ lib/sinon/fake.js
@@ -2,8 +2,6 @@
 
 var arrayProto = require("@sinonjs/commons").prototypes.array;
 var createProxy = require("./proxy");
-var fakeApi = require("./fake-api");
-var extend = require("./util/core/extend");
 var nextTick = require("./util/core/next-tick");
 
 var slice = arrayProto.slice;
@@ -25,7 +23,6 @@ function wrapFunc(f) {
         return f && f.apply(this, arguments);
     };
     proxy = createProxy(fakeInstance, f || fakeInstance);
-    extend(proxy, fakeApi);
 
     proxy.displayName = "fake";
     proxy.id = "fake#" + uuid++;
