+++ lib/sinon/mock-expectation.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var arrayProto = require("@sinonjs/commons").prototypes.array;
-var spyInvoke = require("./spy").invoke;
+var proxyInvoke = require("./proxy-invoke");
 var spyCallToString = require("./call").toString;
 var timesInWords = require("./util/core/times-in-words");
 var extend = require("./util/core/extend");
@@ -80,7 +80,7 @@ var mockExpectation = {
     invoke: function invoke(func, thisValue, args) {
         this.verifyCallAllowed(thisValue, args);
 
-        return spyInvoke.apply(this, arguments);
+        return proxyInvoke.apply(this, arguments);
     },
 
     atLeast: function atLeast(num) {
