+++ lib/sinon/proxy.js
@@ -1,8 +1,115 @@
 "use strict";
 
-var slice = require("@sinonjs/commons").prototypes.array.slice;
+var arrayProto = require("@sinonjs/commons").prototypes.array;
 var extend = require("./util/core/extend");
 var functionToString = require("./util/core/function-to-string");
+var fakeCall = require("./call");
+var callUtil = require("./call-util");
+var proxyInvoke = require("./proxy-invoke");
+
+var push = arrayProto.push;
+var forEach = arrayProto.forEach;
+var slice = arrayProto.slice;
+var filter = Array.prototype.filter;
+
+// Public API
+var proxyApi = {
+    toString: functionToString,
+
+    invoke: proxyInvoke,
+
+    matchingFakes: function(args, strict) {
+        return filter.call(this.fakes, function(fake) {
+            return fake.matches(args, strict);
+        });
+    },
+
+    getCall: function getCall(i) {
+        if (i < 0 || i >= this.callCount) {
+            return null;
+        }
+
+        return fakeCall(
+            this,
+            this.thisValues[i],
+            this.args[i],
+            this.returnValues[i],
+            this.exceptions[i],
+            this.callIds[i],
+            this.errorsWithCallStack[i]
+        );
+    },
+
+    getCalls: function() {
+        var calls = [];
+        var i;
+
+        for (i = 0; i < this.callCount; i++) {
+            push(calls, this.getCall(i));
+        }
+
+        return calls;
+    },
+
+    resetHistory: function() {
+        if (this.invoking) {
+            var err = new Error(
+                "Cannot reset Sinon function while invoking it. " +
+                    "Move the call to .resetHistory outside of the callback."
+            );
+            err.name = "InvalidResetException";
+            throw err;
+        }
+
+        this.called = false;
+        this.notCalled = true;
+        this.calledOnce = false;
+        this.calledTwice = false;
+        this.calledThrice = false;
+        this.callCount = 0;
+        this.firstCall = null;
+        this.secondCall = null;
+        this.thirdCall = null;
+        this.lastCall = null;
+        this.args = [];
+        this.lastArg = null;
+        this.returnValues = [];
+        this.thisValues = [];
+        this.exceptions = [];
+        this.callIds = [];
+        this.errorsWithCallStack = [];
+
+        forEach(this.fakes, function(fake) {
+            fake.resetHistory();
+        });
+
+        return this;
+    }
+};
+
+var delegateToCalls = callUtil.delegateToCalls;
+delegateToCalls(proxyApi, "calledOn", true);
+delegateToCalls(proxyApi, "alwaysCalledOn", false, "calledOn");
+delegateToCalls(proxyApi, "calledWith", true);
+delegateToCalls(proxyApi, "calledOnceWith", true, "calledWith", false, undefined, 1);
+delegateToCalls(proxyApi, "calledWithMatch", true);
+delegateToCalls(proxyApi, "alwaysCalledWith", false, "calledWith");
+delegateToCalls(proxyApi, "alwaysCalledWithMatch", false, "calledWithMatch");
+delegateToCalls(proxyApi, "calledWithExactly", true);
+delegateToCalls(proxyApi, "calledOnceWithExactly", true, "calledWithExactly", false, undefined, 1);
+delegateToCalls(proxyApi, "alwaysCalledWithExactly", false, "calledWithExactly");
+delegateToCalls(proxyApi, "neverCalledWith", false, "notCalledWith", false, function() {
+    return true;
+});
+delegateToCalls(proxyApi, "neverCalledWithMatch", false, "notCalledWithMatch", false, function() {
+    return true;
+});
+delegateToCalls(proxyApi, "threw", true);
+delegateToCalls(proxyApi, "alwaysThrew", false, "threw");
+delegateToCalls(proxyApi, "returned", true);
+delegateToCalls(proxyApi, "alwaysReturned", false, "returned");
+delegateToCalls(proxyApi, "calledWithNew", true);
+delegateToCalls(proxyApi, "alwaysCalledWithNew", false, "calledWithNew");
 
 function createProxy(func, originalFunc) {
     var proxy = wrapFunction(func, originalFunc);
@@ -12,9 +119,7 @@ function createProxy(func, originalFunc) {
 
     proxy.prototype = func.prototype;
 
-    extend.nonEnum(proxy, {
-        toString: functionToString
-    });
+    extend.nonEnum(proxy, proxyApi);
 
     return proxy;
 }
