+++ lib/sinon/spy.js
@@ -7,13 +7,11 @@ var functionName = require("@sinonjs/commons").functionName;
 var getPropertyDescriptor = require("./util/core/get-property-descriptor");
 var deepEqual = require("@sinonjs/samsam").deepEqual;
 var isEsModule = require("./util/core/is-es-module");
-var spyCall = require("./call");
 var callUtil = require("./call-util");
 var walkObject = require("./util/core/walk-object");
 var wrapMethod = require("./util/core/wrap-method");
 var sinonFormat = require("./util/core/format");
 var valueToString = require("@sinonjs/commons").valueToString;
-var fakeApi = require("./fake-api");
 
 /* cache references to library methods so that they also can be stubbed without problems */
 var forEach = arrayProto.forEach;
@@ -22,65 +20,6 @@ var push = arrayProto.push;
 var slice = arrayProto.slice;
 
 var uuid = 0;
-
-function createSpy(func) {
-    var name;
-    var funk = func;
-
-    if (typeof funk !== "function") {
-        funk = function() {
-            return;
-        };
-    } else {
-        name = functionName(funk);
-    }
-
-    var proxy = createProxy(funk, funk);
-
-    // Inherit spy API:
-    extend.nonEnum(proxy, spy);
-    extend.nonEnum(proxy, {
-        displayName: name || "spy",
-        instantiateFake: createSpy,
-        id: "spy#" + uuid++
-    });
-    return proxy;
-}
-
-function spy(object, property, types) {
-    var descriptor, methodDesc;
-
-    if (isEsModule(object)) {
-        throw new TypeError("ES Modules cannot be spied");
-    }
-
-    if (!property && typeof object === "function") {
-        return createSpy(object);
-    }
-
-    if (!property && typeof object === "object") {
-        return walkObject(spy, object);
-    }
-
-    if (!object && !property) {
-        return createSpy(function() {
-            return;
-        });
-    }
-
-    if (!types) {
-        return wrapMethod(object, property, createSpy(object[property]));
-    }
-
-    descriptor = {};
-    methodDesc = getPropertyDescriptor(object, property);
-
-    forEach(types, function(type) {
-        descriptor[type] = createSpy(methodDesc[type]);
-    });
-
-    return wrapMethod(object, property, descriptor);
-}
 
 // Public API
 var spyApi = {
@@ -231,7 +170,64 @@ delegateToCalls(spyApi, "yieldToOn", false, "yieldToOn", true, function(property
 });
 /* eslint-enable local-rules/no-prototype-methods */
 
-extend(spy, fakeApi);
+function createSpy(func) {
+    var name;
+    var funk = func;
+
+    if (typeof funk !== "function") {
+        funk = function() {
+            return;
+        };
+    } else {
+        name = functionName(funk);
+    }
+
+    var proxy = createProxy(funk, funk);
+
+    // Inherit spy API:
+    extend.nonEnum(proxy, spyApi);
+    extend.nonEnum(proxy, {
+        displayName: name || "spy",
+        instantiateFake: createSpy,
+        id: "spy#" + uuid++
+    });
+    return proxy;
+}
+
+function spy(object, property, types) {
+    var descriptor, methodDesc;
+
+    if (isEsModule(object)) {
+        throw new TypeError("ES Modules cannot be spied");
+    }
+
+    if (!property && typeof object === "function") {
+        return createSpy(object);
+    }
+
+    if (!property && typeof object === "object") {
+        return walkObject(spy, object);
+    }
+
+    if (!object && !property) {
+        return createSpy(function() {
+            return;
+        });
+    }
+
+    if (!types) {
+        return wrapMethod(object, property, createSpy(object[property]));
+    }
+
+    descriptor = {};
+    methodDesc = getPropertyDescriptor(object, property);
+
+    forEach(types, function(type) {
+        descriptor[type] = createSpy(methodDesc[type]);
+    });
+
+    return wrapMethod(object, property, descriptor);
+}
+
 extend(spy, spyApi);
-spy.spyCall = spyCall;
 module.exports = spy;
