+++ lib/sinon/stub.js
@@ -164,8 +164,6 @@ var proto = {
         });
     },
 
-    resetHistory: spy.resetHistory,
-
     reset: function() {
         this.resetHistory();
         this.resetBehavior();
